src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
azure-common==1.1.28
    # via azure-mgmt-compute
azure-core==1.32.0
    # via
    #   azure-identity
    #   azure-mgmt-core
azure-identity==1.17.1
    # via az-vmcount-exporter (pyproject.toml)
azure-mgmt-compute==33.0.0
    # via az-vmcount-exporter (pyproject.toml)
azure-mgmt-core==1.5.0
    # via azure-mgmt-compute
certifi==2024.8.30
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.0
    # via requests
cryptography==44.0.0
    # via
    #   azure-identity
    #   msal
    #   pyjwt
idna==3.10
    # via requests
isodate==0.7.2
    # via azure-mgmt-compute
msal==1.31.1
    # via
    #   azure-identity
    #   msal-extensions
msal-extensions==1.2.0
    # via azure-identity
portalocker==2.10.1
    # via msal-extensions
prometheus-client==0.17.1
    # via az-vmcount-exporter (pyproject.toml)
pycparser==2.22
    # via cffi
pyjwt==2.10.1
    # via msal
python-dotenv==1.0.0
    # via az-vmcount-exporter (pyproject.toml)
requests==2.32.3
    # via
    #   azure-core
    #   msal
six==1.16.0
    # via azure-core
typing-extensions==4.12.2
    # via
    #   azure-core
    #   azure-identity
    #   azure-mgmt-compute
urllib3==2.2.3
    # via requests
